refactor: reorganize and clean up imports across components
- Moved `editor` to `ui` folder and updated related imports.
- Removed unnecessary `"use client"` directives from components.
- Standardized import file extensions to `.tsx` for consistency.
