+++ src/components/ui/rating-group-advanced.tsx
@@ -1,9 +1,7 @@
-"use client";
-
 import { HeartIcon, StarIcon } from "lucide-react";
 import * as React from "react";
-import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
-import { cn } from "@/lib/utils";
+import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group.tsx";
+import { cn } from "@/lib/utils.ts";
 
 interface RatingGroupAdvancedProps {
   value?: string;
