+++ src/components/ui/toggle-group.tsx
@@ -1,11 +1,8 @@
-"use client";
-
-import * as React from "react";
 import { type VariantProps } from "class-variance-authority";
 import { ToggleGroup as ToggleGroupPrimitive } from "radix-ui";
-
-import { cn } from "@/lib/utils";
+import * as React from "react";
 import { toggleVariants } from "@/components/ui/toggle";
+import { cn } from "@/lib/utils";
 
 const ToggleGroupContext = React.createContext<
   VariantProps<typeof toggleVariants> & {
